requirements.txt
absl-py==0.8.1
aiofiles==23.2.1
aiohttp==3.8.3
aiosignal==1.2.0
alembic==1.8.1
altair==4.2.0
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
aspy.yaml==1.3.0
astor==0.8.1
astroid==2.4.1
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
audioread==2.1.8
autopep8==1.6.0
Babel==2.8.0
backcall==0.1.0
backports.zoneinfo==0.2.1
bce-python-sdk==0.8.53
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
bokeh==2.4.3
boltons==23.1.1
cachetools==4.0.0
certifi==2019.9.11
certipy==0.1.3
cffi==1.15.1
cfgv==2.0.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.0.4
cloudpickle==1.6.0
cma==2.7.0
colorama==0.4.4
colorlog==4.1.0
commonmark==0.9.1
cryptography==38.0.1
cycler==0.10.0
Cython==0.29
datasets==2.7.0
debugpy==1.6.0
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.3
easydict==1.9
einops==0.6.1
entrypoints==0.4
et-xmlfile==1.0.1
exceptiongroup==1.1.3
fastapi==0.85.1
fastjsonschema==2.16.1
ffmpy==0.3.1
filelock==3.0.12
fire==0.5.0
flake8==4.0.1
Flask==1.1.1
Flask-Babel==1.0.0
Flask-Cors==3.0.8
forbiddenfruit==0.1.3
frozenlist==1.3.0
fsspec==2022.11.0
ftfy==6.1.1
funcsigs==1.0.2
future==0.18.0
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.14
google-auth==1.10.0
google-auth-oauthlib==0.4.1
gradio==3.34.0
gradio_client==0.2.6
graphviz==0.13
greenlet==1.1.3
grpcio==1.35.0
gunicorn==20.0.4
gym==0.12.1
h11==0.14.0
h5py==2.9.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
identify==1.4.10
idna==2.8
imageio==2.6.1
imageio-ffmpeg==0.3.0
importlib-metadata==4.2.0
importlib-resources==5.9.0
intervaltree==3.1.0
ipykernel @ file:///tmp/output/deps/ipykernel-6.9.1-py3-none-any.whl
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
jieba==0.42.1
Jinja2==3.0.0
joblib==0.14.1
JPype1==0.7.2
json5==0.9.5
jsonschema==4.16.0
jupyter-archive==3.2.1
jupyter-core==4.11.1
jupyter-lsp==1.5.1
jupyter-server @ file:///tmp/output/deps/jupyter_server-1.16.0-py3-none-any.whl
jupyter-telemetry==0.1.0
jupyter_client==7.3.5
jupyterhub==1.3.0
jupyterlab==3.4.5
jupyterlab-language-pack-zh-CN==3.4.post1
jupyterlab-pygments==0.2.2
jupyterlab-server==2.10.3
jupyterlab-widgets==3.0.3
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
librosa==0.7.2
lightgbm==3.1.1
linkify-it-py==2.0.2
llvmlite==0.31.0
lxml==4.9.1
Mako==1.2.2
Markdown==3.1.1
markdown-it-py==2.2.0
MarkupSafe==2.0.1
matplotlib==2.2.3
matplotlib-inline==0.1.6
mccabe==0.6.1
mdit-py-plugins==0.3.3
mdurl==0.1.1
mido==1.3.0
mistune==0.8.4
more-itertools==7.2.0
moviepy==1.0.1
multidict==6.0.2
multiprocess==0.70.11.1
nbclassic==0.3.1
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.5.0
nest-asyncio==1.5.5
netifaces==0.10.9
networkx==2.4
nltk==3.4.5
nodeenv==1.3.4
note-seq==0.0.3
notebook==5.7.0
numba==0.48.0
numpy==1.19.5
oauthlib==3.1.0
objgraph==3.4.1
omegaconf==2.3.0
onnx==1.14.1
opencv-python==4.6.0.66
openpyxl==3.0.5
opt-einsum==3.3.0
orjson==3.9.7
packaging==23.2
paddle-bfloat==0.1.7
paddle2onnx==1.0.0
paddlefsl==1.1.0
paddlehub==2.3.0
paddlenlp==2.5.1
paddlepaddle-gpu @ file:///tmp/paddlepaddle_gpu-2.4.0.post112-cp37-cp37m-linux_x86_64.whl
paddlesde==0.2.5
pamela==1.0.0
pandas==1.1.5
pandocfilters==1.5.0
parameterized==0.9.0
parl==1.4.1
parso==0.7.1
pathlib==1.0.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==8.2.0
pipreqs==0.4.13
pkgutil_resolve_name==1.3.10
plotly==5.8.0
pluggy==1.0.0
ppdiffusers==0.11.0
pre-commit==1.21.0
pretty-midi==0.2.10
prettytable==0.7.2
proglog==0.1.9
prometheus-client==0.14.1
prompt-toolkit==2.0.10
protobuf==3.20.2
psutil==5.7.2
ptyprocess==0.7.0
py4j==0.10.9.2
pyarrow==10.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pybboxes==0.1.1
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.9.9
pydantic==1.10.13
pydantic_core==2.10.1
pydeck==0.8.0
pydocstyle==5.0.2
pydub==0.25.1
pyflakes==2.4.0
pyglet==1.4.5
Pygments==2.13.0
pyhumps==3.8.0
pylint==2.5.2
Pympler==1.0.1
pynvml==8.0.4
pyOpenSSL==22.0.0
pyparsing==3.0.9
pypmml==0.9.11
pyrsistent==0.18.1
python-dateutil==2.8.2
python-json-logger==2.0.4
python-jsonrpc-server==0.3.4
python-language-server==0.33.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.5.0
python-multipart==0.0.6
pytz==2019.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.1.2
pyzmq==23.2.1
rarfile==3.1
recordio==0.1.7
regex==2023.10.3
requests==2.24.0
requests-mock==1.11.0
requests-oauthlib==1.3.0
resampy==0.2.2
responses==0.18.0
rich==12.6.0
rope==0.17.0
rsa==4.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
safetensors==0.2.8
sahi==0.10.1
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.10.0
semantic-version==2.10.0
semver==2.13.0
Send2Trash==1.8.0
sentencepiece==0.1.96
seqeval==1.2.2
shapely==2.0.0
shellcheck-py==0.7.1.1
simplegeneric==0.8.1
six==1.16.0
sklearn==0.0
smmap==3.0.5
sniffio==1.3.0
snowballstemmer==2.0.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
starlette==0.20.4
streamlit==1.13.0
streamlit-image-comparison==0.0.3
tabulate==0.8.3
tb-nightly==1.15.0a20190801
tb-paddle==0.3.6
tenacity==8.0.1
tensorboard==2.1.0
tensorboardX==1.8
termcolor==1.1.0
terminado==0.15.0
terminaltables==3.1.10
testpath==0.4.2
threadpoolctl==2.1.0
tinycss2==1.1.1
toml==0.10.0
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
trampoline==0.1.2
typed-ast==1.4.1
typer==0.9.0
typing_extensions==4.7.1
tzdata==2022.7
tzlocal==4.2
uc-micro-py==1.0.2
ujson==1.35
urllib3==1.25.11
uvicorn==0.22.0
validators==0.20.0
virtualenv==16.7.9
visualdl==2.4.0
watchdog==2.2.0
wcwidth==0.2.9
webencodings==0.5.1
websocket-client==1.4.1
websockets==11.0.3
Werkzeug==0.16.0
whatthepatch==1.0.2
widgetsnbextension==3.5.2
wrapt==1.12.1
xarray==0.16.2
xgboost==1.3.3
xlrd==1.2.0
xxhash==3.1.0
yapf==0.26.0
yarg==0.1.9
yarl==1.7.2
zipp==3.8.1
